+++ services/data_integration/src/oci_cli_data_integration/dataintegration_cli_extended.py
@@ -13,7 +13,7 @@ dataintegration_cli.connection_group.commands.pop(dataintegration_cli.create_con
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.create_connection_create_connection_from_atp.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.create_connection_create_connection_from_adwc.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.create_connection_create_connection_from_oracle.name)
-dataintegration_cli.connection_group.commands.pop(dataintegration_cli.create_connection_create_connection_from_lakehouse.name)
+dataintegration_cli.connection_group.commands.pop(dataintegration_cli.create_connection_create_connection_from_lake.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.create_connection_create_connection_from_rest_basic_auth.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.create_connection_create_connection_from_rest_no_auth.name)
 
@@ -115,7 +115,7 @@ def create_connection_extended(ctx, **kwargs):
                 'password': kwargs['password']
             }
         )
-        ctx.invoke(dataintegration_cli.create_connection_create_connection_from_lakehouse, **oracle_db_args)
+        ctx.invoke(dataintegration_cli.create_connection_create_connection_from_lake, **oracle_db_args)
     elif 'model_type' in kwargs and kwargs['model_type'] == 'ORACLE_REST_BASIC_AUTH_CONNECTION':
         oracle_rest_auth_args = {}
         oracle_rest_auth_args.update(common_args)
@@ -142,7 +142,7 @@ dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_con
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_connection_update_connection_from_atp.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_connection_update_connection_from_adwc.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_connection_update_connection_from_oracle.name)
-dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_connection_update_connection_from_lakehouse.name)
+dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_connection_update_connection_from_lake.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_connection_update_connection_from_rest_basic_auth.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_connection_update_connection_from_rest_no_auth.name)
 
@@ -248,7 +248,7 @@ def update_connection_extended(ctx, **kwargs):
                 'password': kwargs['password']
             }
         )
-        ctx.invoke(dataintegration_cli.update_connection_update_connection_from_lakehouse, **oracle_db_args)
+        ctx.invoke(dataintegration_cli.update_connection_update_connection_from_lake, **oracle_db_args)
     elif 'model_type' in kwargs and kwargs['model_type'] == 'ORACLE_REST_BASIC_AUTH_CONNECTION':
         oracle_rest_auth_args = {}
         oracle_rest_auth_args.update(common_args)
@@ -1091,7 +1091,7 @@ dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_data_asset_from_adwc.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_data_asset_from_atp.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_data_asset_from_object_storage.name)
-dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_data_asset_from_lakehouse.name)
+dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_data_asset_from_lake.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_data_asset_from_rest.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_connection_from_atp.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_connection_from_adwc.name)
@@ -1099,7 +1099,7 @@ dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_connection_from_object_storage.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_connection_from_my_sql.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_connection_from_jdbc.name)
-dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_connection_from_lakehouse.name)
+dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_connection_from_lake.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_connection_from_rest_basic_auth.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_connection_from_rest_no_auth.name)
 dataintegration_cli.connection_validation_group.commands.pop(dataintegration_cli.create_connection_validation_create_data_asset_from_my_sql.name)
@@ -1110,11 +1110,11 @@ dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_con
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_connection_update_connection_from_my_sql.name)
 dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.create_data_asset_create_data_asset_from_jdbc.name)
 dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.create_data_asset_create_data_asset_from_my_sql.name)
-dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.create_data_asset_create_data_asset_from_lakehouse.name)
+dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.create_data_asset_create_data_asset_from_lake.name)
 dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.create_data_asset_create_data_asset_from_rest.name)
 dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.update_data_asset_update_data_asset_from_jdbc.name)
 dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.update_data_asset_update_data_asset_from_my_sql.name)
-dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.update_data_asset_update_data_asset_from_lakehouse.name)
+dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.update_data_asset_update_data_asset_from_lake.name)
 dataintegration_cli.data_asset_group.commands.pop(dataintegration_cli.update_data_asset_update_data_asset_from_rest.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.create_connection_create_connection_from_amazon_s3.name)
 dataintegration_cli.connection_group.commands.pop(dataintegration_cli.update_connection_update_connection_from_amazon_s3.name)